pyproject.toml
[project]
name = "from-scratch-nn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.5",
    "pillow>=11.2.1",
]

[dependency-groups]
test = [
    "pytest>=8.3.5",
]
